refactor: 검증을 위해 높이 클래스 분리
