+++ src/main/java/LadderController.java
@@ -1,3 +1,4 @@
+import domain.Height;
 import domain.Ladder;
 import domain.Names;
 import view.InputView;
@@ -9,16 +10,20 @@ public class LadderController {
 
     public static void main(String[] args) {
         final Names names = retryOnException(LadderController::getNames);
-        final int height = retryOnException(InputView::readHeight);
+        final Height height = retryOnException(LadderController::getHeight);
 
         final int width = names.names().size() - 1;
-        final Ladder ladder = new Ladder(width, height);
+        final Ladder ladder = new Ladder(width, height.height());
 
         OutputView.printResult(names, ladder);
     }
 
     private static Names getNames() {
         return new Names(InputView.readNames());
+    }
+
+    private static Height getHeight() {
+        return new Height(InputView.readHeight());
     }
 
     private static <T> T retryOnException(Supplier<T> supplier) {
