+++ src/main/java/domain/Ladder.java
@@ -5,13 +5,10 @@ import java.util.stream.IntStream;
 
 
 public class Ladder {
-    private static final int MIN_HEIGHT = 1;
-    private static final int MAX_HEIGHT = 10;
 
     private final List<Line> lines;
 
     public Ladder(final int width, final int height) {
-        validateHeight(height);
         this.lines = generateLadder(width, height);
     }
 
@@ -23,12 +20,5 @@ public class Ladder {
 
     public List<Line> getLines() {
         return this.lines;
-    }
-
-    private void validateHeight(final int height) {
-        final String errorMessage = String.format("높이는 %d 이상 %d 이하로 입력해 주세요.", MIN_HEIGHT, MAX_HEIGHT);
-        if (MIN_HEIGHT > height || height > MAX_HEIGHT) {
-            throw new IllegalArgumentException(errorMessage);
-        }
     }
 }
